requirements.txt
streamlit==1.39.0
pandas==2.2.2
plotly
python-dotenv
pillow
msal